% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joycon_list.R
\docType{data}
\name{joycon_list}
\alias{joycon_list}
\title{List of Nintendo-inspired palettes available.}
\format{
An object of class \code{character} of length 47.
}
\usage{
joycon_list
}
\description{
Palettes are available in three, five, and eight colors.

Use \code{joycon_list} to see all palettes available.
}
\keyword{datasets}
