% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{joycon_palettes}
\alias{joycon_palettes}
\title{Palettes inspired by Nintendo.}
\format{
An object of class \code{list} of length 19.
}
\usage{
joycon_palettes
}
\description{
Color options include: \code{supermario}, \code{skywardsword}, \code{starbits},
\code{newhorizons}, \code{rainbowroad}, \code{splatoon},
\code{yoshi}, \code{isabelle}, \code{kirby}, \code{twilightprincess}, \code{wiisports},
\code{korok}, \code{majora}, \code{duckhunt}, \code{littlemac},
\code{samus}, \code{delfino}, and \code{hyrule}.
}
\keyword{datasets}
