% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joycon.R
\docType{package}
\name{joycon}
\alias{joycon}
\title{joycon}
\description{
The \code{joycon} package provides several color palettes inspired by Nintendo games,
including Super Mario, Animal Crossing, and the Legend of Zelda.
\if{html}{\figure{hexsticker.PNG}{options: width="20\%"}}
}
\author{
Julianna Calabrese
}
\keyword{color}
\keyword{nintendo}
\keyword{palette}
