% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{joycon_pal}
\alias{joycon_pal}
\title{An Nintendo-inspired color palette generator}
\usage{
joycon_pal(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of color palette.
Palettes include: \code{supermario}, \code{skywardsword}, \code{starbits},
\code{newhorizons}, \code{rainbowroad}, \code{splatoon},  \code{princesspeach},
\code{yoshi}, \code{isabelle}, \code{kirby}, \code{twilightprincess}, \code{wiisports},
\code{korok}, \code{majora}, \code{duckhunt}, \code{littlemac},
\code{samus}, \code{delfino}, and \code{hyrule}.}

\item{n}{Number of colors needed.}

\item{type}{Continuous or discrete.}
}
\description{
This includes palettes inspired by Nintendo games. If you have a specific game you'd like to request, DM me on Twitter @jrosecalabrese.
}
