% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{joycon_pal}
\alias{joycon_pal}
\title{An Nintendo-inspired color palette generator}
\usage{
joycon_pal(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of color palette. Please see \code{joycon_palettes} for full list.}

\item{n}{Number of colors needed.}

\item{type}{Continuous or discrete.}
}
\description{
This includes palettes inspired by Nintendo games.
If you have a specific game you'd like to request, DM me on Twitter @jrosecalabrese.
}
