% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{joycon_palettes}
\alias{joycon_palettes}
\title{Look at these nice colors.}
\format{
An object of class \code{list} of length 7.
}
\usage{
joycon_palettes
}
\description{
I need to fix that function and add an actual description later.
}
\keyword{datasets}
