% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joycon_palettes.R
\docType{data}
\name{joycon_palettes}
\alias{joycon_palettes}
\title{Palettes inspired by Nintendo.}
\format{
An object of class \code{list} of length 47.
}
\usage{
joycon_palettes
}
\description{
Color palettes inspired by Nintendo.
Palettes are available in three, five, and eight colors.

Use \code{joycon_palettes_list} to see all palettes available.
}
\keyword{datasets}
